genre/requirements.txt
librosa>=0.8.0
tensorflow==2.14.0
numpy>=1.14.5,<1.24
urllib3<=2.0.7
protobuf>=3.19.6,<4.24
boto3
pyyaml